% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testmaker_df_colcontents.R
\name{dput_to_string}
\alias{dput_to_string}
\title{Helper function to convert string to R code to regenerate that string.}
\usage{
dput_to_string(object)
}
\arguments{
\item{object}{R object, expecting a vector}
}
\value{
Character atomic of R code to recreate the object. Plays well with \code{glue} functions and \code{cat()}.
}
\description{
Helper function to convert string to R code to regenerate that string.
}
\examples{
\dontrun{
temp = dput_to_string(rownames(mtcars))
cat(temp)
}
}
