% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_testmaker.R
\name{finish_testmaker}
\alias{finish_testmaker}
\title{Handle the standardized return options for all stopifnotmaker and testmaker functions}
\usage{
finish_testmaker(test.text, return.style, silent)
}
\arguments{
\item{test.text}{Character vector of the finished lines of codes to display or provide as appropriate.}

\item{return.style}{Designation for what to return. "clip" returns nothing, but loads the text into the clipboard.
"text" returns the text as a character vector, "none" returns nothing}

\item{silent}{Suppress printing text to console? logical; default of false.}
}
\value{
Either nothing or a character vector of lines of R code for writing a testthat test, depending on \code{return.style}.
}
\description{
Internal helper function.
}
