% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_testmaker.R
\name{validate_testmaker}
\alias{validate_testmaker}
\title{Validation function for testmaker functions}
\usage{
validate_testmaker(
  x,
  return.style,
  silent,
  object.name = NULL,
  call = rlang::caller_env()
)
}
\arguments{
\item{x}{Template dataframe of target object (e.g. we expect our test object to share properties with this)}

\item{return.style}{Defines what is returned by testmaker functions. SHOULD be "clip", "text", or "none"}

\item{silent}{Suppress printing text to console? logical; default of false.}

\item{object.name}{Name of the object to apply the stopifnot to; presumably the name of the dataframe argument
in the function the test is being written for. Defaults to NULL; provide when validating a stopifnot_* function.}
}
\description{
Internal helper function to streamline validation
}
